pyproject.toml
[project]
name = "slack-bot-template"
version = "0.1.0"
description = "Template for a Slack bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "slack-bolt>=1.23.0",
]
